pyproject.toml
[project]
name = "fidcsv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "charapi",
    "python-box[yaml]>=7.3.2",
]

[project.scripts]
fidcsv = "fidcsv.main:main"

[tool.uv]
package = true

[tool.uv.sources]
charapi = { path = "../charapi", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
